Memory allocation cleanup and integer stringification.
Added ImportedFunction class, which also stores its import name,
and uses it when its import method called.

Added helper C functions to manage arrays.

Reference counts and array lengths are now properly accessed
as long, not int.

Added Integer stringify method as an imported function.

+++ utf8.c
@@ -1,5 +1,5 @@
 
-int decode_utf8_raw(const char *bytes, int byte_length, unsigned short *characters) {
+long decode_utf8_raw(const char *bytes, long byte_length, unsigned short *characters) {
     const char *bytes_end = bytes + byte_length;
     const unsigned short *characters_start = characters;
     
@@ -53,7 +53,7 @@ int decode_utf8_raw(const char *bytes, int byte_length, unsigned short *characte
 }
 
 
-int encode_utf8_raw(const unsigned short *characters, int character_length, char *bytes) {
+long encode_utf8_raw(const unsigned short *characters, long character_length, char *bytes) {
     const unsigned short *characters_end = characters + character_length;
     const char *bytes_start = bytes;
